# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Boxty API
  version: 1.0.0
  description: |
    Boxty — programmable sandbox platform. Run code in cloud with gVisor isolation,
    secrets, cron scheduling, image building, persistent volumes, and web endpoint API.
servers:
  - url: https://api.boxty.dev
    description: Production
  - url: http://localhost:3000
    description: Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from login/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key format "Bearer bxty_..."
  schemas:
...
    ImageBuild:
      type: object
      properties:
        id: { type: string }
        image_url: { type: string, nullable: true }
        status: { type: string, enum: [building, done, failed] }
        error: { type: string, nullable: true }
...
    Deployment:
      type: object
      properties:
        id: { type: string }
        url: { type: string, format: uri }
        status: { type: string, enum: [deploying, active, failed] }
...
    Environment:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, enum: [development, staging, production] }
        workspace_id: { type: string }
    App:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }
        image: { type: string }
        cpu: { type: integer }
        memory: { type: integer }
        gpu: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    UsageReport:
      type: object
      properties:
        total_cost: { type: number }
...
paths:
...
  /api/images/build:
    post:
      summary: Build image
      tags: [Images]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [base_image, commands]
              properties:
                base_image: { type: string }
                commands: { type: array, items: { type: string } }
                name: { type: string }
      responses:
        '201':
          description: Build started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageBuild' }
  /api/images/build-from-dockerfile:
    post:
      summary: Build from Dockerfile
      tags: [Images]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                dockerfile: { type: string, format: binary }
                context: { type: string, format: binary }
      responses:
        '201':
          description: Build started
  /api/images:
    get:
      summary: List images
      tags: [Images]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Images list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ImageBuild' }
  /api/images/{id}:
    get:
      summary: Get image status
      tags: [Images]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Image details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageBuild' }
    delete:
      summary: Delete image
      tags: [Images]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Image deleted
...
  /api/deployments:
    post:
      summary: Create deployment
      tags: [Deployments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, image]
              properties:
                name: { type: string }
                image: { type: string }
                cpu: { type: integer }
                memory: { type: integer }
                gpu: { type: string }
      responses:
        '201':
          description: Deployment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
    get:
      summary: List deployments
      tags: [Deployments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Deployments list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Deployment' }
  /api/deployments/{id}:
    get:
      summary: Get deployment
      tags: [Deployments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Deployment' }
    delete:
      summary: Delete deployment
      tags: [Deployments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deployment deleted
  /api/deployments/{id}/invoke:
    post:
      summary: Invoke deployment
      tags: [Deployments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Invocation result
...
  /api/environments:
    post:
      summary: Create environment
      tags: [Environments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, name]
              properties:
                workspace_id: { type: string }
                name: { type: string }
                type: { type: string, enum: [development, staging, production] }
      responses:
        '201':
          description: Environment created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }
    get:
      summary: List environments
      tags: [Environments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Environments list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Environment' }
  /api/environments/{id}:
    delete:
      summary: Delete environment
      tags: [Environments]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Environment deleted
  /api/apps:
    post:
      summary: Create app
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, env_id, name, image, cpu, memory]
              properties:
                workspace_id: { type: string }
                env_id: { type: string }
                name: { type: string }
                image: { type: string }
                cpu: { type: integer }
                memory: { type: integer }
                gpu: { type: string }
                timeout: { type: integer }
      responses:
        '201':
          description: App created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/App' }
    get:
      summary: List apps
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      responses:
        '200':
          description: Apps list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/App' }
  /api/apps/{id}:
    get:
      summary: Get app
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: App details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/App' }
    delete:
      summary: Delete app
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: App deleted
  /api/apps/{id}/stop:
    post:
      summary: Stop app
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: App stopped
  /api/apps/{id}/deploy:
    post:
      summary: Deploy app
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image: { type: string }
                cpu: { type: integer }
                memory: { type: integer }
                gpu: { type: string }
      responses:
        '200':
          description: App deployed
  /api/apps/{id}/sandboxes:
    get:
      summary: Get app sandboxes
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sandboxes list
  /api/apps/{id}/deployments:
    get:
      summary: Get app deployments
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deployments list
...
  /api/apps/{id}/usage:
    get:
      summary: Get app usage
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: App usage
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageReport' }
  /api/apps/{id}/logs:
    get:
      summary: Get app logs
      tags: [Apps]
      security: [{ BearerAuth: [] }, { ApiKeyAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: App logs